vineflower 1.11.1 (new formula)
Vineflower is a popular Java decompiler based on JetBrains' Fernflower.
This attempt at a formula for Vineflower 1.11.1 downloads the source
code from the appropriate GitHub release and builds it using the
project's Gradle build task. Because two Gradle tests appear to fail in
this version of the repository, they are skipped after building. The
resulting binary is functional.

A short test block is included that checks Vineflower against a basic
Java class. Vineflower can yield a successful exit code even if it
produces an error. Because of this, the script checks that the word
'error' is not included in the output.